rename sentinal apache dubbo package name
