+++ sentinel-adapter/sentinel-apache-dubbo-adapter/src/main/java/com/alibaba/csp/sentinel/adapter/apache/dubbo/fallback/DubboFallback.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.csp.sentinel.adapter.dubbo.fallback;
+package com.alibaba.csp.sentinel.adapter.apache.dubbo.fallback;
 
 import com.alibaba.csp.sentinel.slots.block.BlockException;
