+++ sentinel-adapter/sentinel-apache-dubbo-adapter/src/test/java/com/alibaba/csp/sentinel/adapter/dubbo/DubboAppContextFilterTest.java
@@ -17,6 +17,8 @@ package com.alibaba.csp.sentinel.adapter.dubbo;
 
 import com.alibaba.csp.sentinel.BaseTest;
 
+import com.alibaba.csp.sentinel.adapter.apache.dubbo.DubboAppContextFilter;
+import com.alibaba.csp.sentinel.adapter.apache.dubbo.DubboUtils;
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.rpc.Invocation;
 import org.apache.dubbo.rpc.Invoker;
